==> notebooks/01_data_ingestion.ipynb <==
import pandas as pd

df = pd.read_parquet("data/raw/yellow_tripdata_2025-01.parquet")
df.head()

df = pd.read_parquet("../Data/Raw/yellow_tripdata_2025-01.parquet")
df.head()


! pip install pandas pyarrow fastparquet

py --version


# This will print the Python version
import sys
print(sys.version)

import pandas as pd

df = pd.read_parquet("data/raw/yellow_tripdata_2025-01.parquet")
df.head()

import pandas as pd

df = pd.read_parquet("../Data/Raw/yellow_tripdata_2025-01.parquet")
df.head()

 ¡ pip install pandas pyarrow fastparquet

pip install pandas pyarrow fastparquet


import os
os.getcwd()

df = pd.read_parquet("../Data/Raw/yellow_tripdata_2025-01.parquet")
df.head()

df = pd.read_parquet("../Data/Raw/yellow_tripdata_2025-01.parquet")
df.head()

import pandas as pd


df = pd.read_parquet("../Data/Raw/yellow_tripdata_2025-01.parquet")
df.head()
