==> notebooks/02_big_data_eda.ipynb <==
# =========================
# IMPORTS
# =========================
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

# =========================
# RUTAS
# =========================
RAW_PATH = "../Data/Raw/yellow_tripdata_2025-01.parquet"
PROCESSED_DIR = "../Data/processed"
REPORTS_DIR = "../reports"

os.makedirs(PROCESSED_DIR, exist_ok=True)
os.makedirs(REPORTS_DIR, exist_ok=True)

print("Rutas configuradas:")
print("RAW_PATH      =", RAW_PATH)
print("PROCESSED_DIR =", PROCESSED_DIR)
print("REPORTS_DIR   =", REPORTS_DIR)

# =========================
# VALIDACIONES
# =========================
print("\nValidación de rutas:")
print("Processed exists:", os.path.exists(PROCESSED_DIR))
print("Reports exists:", os.path.exists(REPORTS_DIR))
print("Raw file exists:", os.path.exists(RAW_PATH))

print("\nCurrent Working Directory (CWD):")
print(os.getcwd())

# =========================
# CARGA DE DATOS
# =========================
cols = [
    "tpep_pickup_datetime",
    "tpep_dropoff_datetime",
    "trip_distance",
    "fare_amount",
    "total_amount",
    "passenger_count",
    "payment_type"
]

df = pd.read_parquet(RAW_PATH, columns=cols)

# =========================
# EXPLORACIÓN BÁSICA
# =========================
print("\nPrimeras filas:")
display(df.head())

print("\nShape del DataFrame:")
print(df.shape)

print("\nInfo del DataFrame:")
df.info()
